Refactor page validation into PageModel
The PageService seemed to have too much domain knowledge, so this
refactors the domain-specific details of page creation and validation
into the page's model itself.
